chore(wmg): refactor pipeline v2
