--- backend/wmg/pipeline/errors.py
@@ -1,2 +1,0 @@
-class UnsupportedMatrixTypeError(Exception):
-    pass
